+++ app/libraries/Jentil.php
@@ -106,6 +106,7 @@ final class Jentil extends AbstractTheme
         $this->setups['Styles\Normalize'] = new Setups\Styles\Normalize($this);
         $this->setups['Styles\Posts'] = new Setups\Styles\Posts($this);
         $this->setups['Styles\Core'] = new Setups\Styles\Core($this);
+        $this->setups['Styles\Gutenberg'] = new Setups\Styles\Gutenberg($this);
     }
 
     private function setUpScripts()